refactor: ItemRepository 설계 변경
